+++ src/mruby_integration/models/mouse.cpp
@@ -60,6 +60,17 @@ mrb_Mouse_position(mrb_state* mrb, mrb_value) -> mrb_value
   return obj;
 }
 
+auto
+mrb_Mouse_set_position(mrb_state* mrb, mrb_value) -> mrb_value
+{
+  Vector2* vector;
+  mrb_get_args(mrb, "d", &vector, &Vector2_type);
+
+  SetMousePosition(vector->x, vector->y);
+
+  return mrb_nil_value();
+}
+
 void
 append_models_Mouse(mrb_state* mrb)
 {
@@ -76,6 +87,8 @@ append_models_Mouse(mrb_state* mrb)
     mrb, Mouse_class, "up?", mrb_Mouse_up, MRB_ARGS_REQ(1));
   mrb_define_class_method(
     mrb, Mouse_class, "position", mrb_Mouse_position, MRB_ARGS_NONE());
+  mrb_define_class_method(
+    mrb, Mouse_class, "position=", mrb_Mouse_set_position, MRB_ARGS_REQ(1));
 
   load_ruby_models_mouse(mrb);
 }